Added barrelling for top bar
